// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/ReporteArticuloMedidas/IArticulosMasterParametros.ts
export interface IArticulosMasterParametros {
        TipoDeConsulta: string;
        Ancho: number;
        AnchoMin: number;
        AnchoMax: number;
        Largo: number;
        LargoMin: number;
        LargoMax: number;
        TipoIndustria: string
        FechaIni:string;
        FechaFin:string;
        OrderByfiltro1: string;
        OrderByfiltro2: string;
        OrderByfiltro3: string;
        OrderBy: string;
    }
    
  